% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc_licor.R
\name{calc_licor}
\alias{calc_licor}
\title{Calculate metrics for LI-COR 6400}
\usage{
calc_licor(licor)
}
\arguments{
\item{licor}{licor object}
}
\description{
Calculate metrics for LI-COR 6400
}
\details{
All calculations are from the formulae in the excel file LI-COR6400 exports
}
\examples{
library("dplyr")
file <- system.file("0000-AAA8891-210718", package = "LICOR6400")
lic <- read_licor6400(file = file)
#add total leaf area - normally use a join
lic <- lic \%>\% mutate(total.leaf.area = 6)
calc_licor(licor = lic)
}
