% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_licor6400.R
\name{read_licor6400}
\alias{read_licor6400}
\title{Read LI-COR 6400 data}
\usage{
read_licor6400(file)
}
\arguments{
\item{file}{Path to file to import}
}
\description{
Read LI-COR 6400 data
}
\examples{
file <- system.file("0000-AAA8891-210718", package = "LICOR6400")
lic <- read_licor6400(file = file)
}
